Source code update for 1.7.1
1. several issues fixed.
2. language update.
3. interface adjustment.
4. update icons and running background images.

+++ safeslingermessager/safeslingermessager/IdleHandler.m
@@ -60,7 +60,7 @@
     }
     
     //convert the wait period into minutes rather than seconds
-    int timeout = [[NSUserDefaults standardUserDefaults]integerForKey:kPasshpraseCacheTime];
+    NSInteger timeout = [[NSUserDefaults standardUserDefaults]integerForKey:kPasshpraseCacheTime];
     if(timeout>0)
         idleTimer = [NSTimer scheduledTimerWithTimeInterval:timeout target:self selector:@selector(idleTimerExceeded) userInfo:nil repeats:NO];
 }
